adds TableTag component
