+++ frontend/src/pages/agreements/list/AgreementsList.scss
@@ -3,8 +3,3 @@
 .usa-table th {
     padding: 0.5rem 0.5rem;
 }
-
-.underReview {
-    background-color: #8B4687;
-    color: white;
-}
